+++ packages/auto_animated/example/lib/screens/animate_on_visibility.dart
@@ -5,9 +5,6 @@ import 'package:auto_animated/auto_animated.dart';
 class AnimateOnVisibilityExample extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final textStyle =
-        Theme.of(context).textTheme.title.copyWith(color: Colors.black);
-
     return Scaffold(
       body: SafeArea(
         // Wrapper before Scroll view!
@@ -16,20 +13,19 @@ class AnimateOnVisibilityExample extends StatelessWidget {
           child: SingleChildScrollView(
             child: Column(
               children: <Widget>[
-                SizedBox(height: 16),
-                Text('Animate elements on visibility', style: textStyle),
                 for (int i = 0; i < 20; i++)
                   AnimateOnVisibilityChange(
                     key: Key('$i'),
                     builder: animationBuilder(
                       SizedBox(
                         width: double.infinity,
-                        height: 200,
+                        height: 128,
                         child: HorizontalItem(
                           title: '$i',
                         ),
                       ),
                       xOffset: i.isEven ? 0.15 : -0.15,
+                      padding: EdgeInsets.all(16),
                     ),
                   ),
               ],