+++ packages/auto_animated/lib/auto_animated.dart
@@ -1,8 +1,7 @@
 export 'src/grid.dart';
+export 'src/helpers/callbacks.dart';
 export 'src/icon_button.dart';
 export 'src/list.dart';
-export 'src/list_animation.dart'
-    show AutoAnimatedListItemBuilder, AutoAnimatedListRemovedItemBuilder;
 export 'src/on_visibility_change.dart';
 export 'src/sliver_grid.dart';
 export 'src/sliver_list.dart';