+++ packages/auto_animated/lib/src/sliver_list.dart
@@ -1,14 +1,13 @@
-import 'dart:async';
-
+import 'package:auto_animated/src/on_visibility_change.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
-import 'list_animation.dart';
+import 'helpers/callbacks.dart';
+import 'helpers/utils.dart' as utils;
 
 const Duration _kDuration = Duration(milliseconds: 300);
 
 class AutoAnimatedSliverList extends StatefulWidget {
-  ///  when they are inserted or removed.
   const AutoAnimatedSliverList({
     @required this.itemBuilder,
     @required this.itemCount,
@@ -31,10 +30,6 @@ class AutoAnimatedSliverList extends StatefulWidget {
   final Duration showItemDuration;
 
   /// Called, as needed, to build list item widgets.
-  ///
-  /// List items are only built when they're scrolled into view.
-  /// Implementations of this callback should assume that
-  /// [AutoAnimatedListState.removeItem] removes an item immediately.
   final AutoAnimatedListItemBuilder itemBuilder;
 
   /// The number of items the list will start with.
@@ -46,15 +41,6 @@ class AutoAnimatedSliverList extends StatefulWidget {
 
   /// Whether the scroll view scrolls in the reading direction.
   ///
-  /// For example, if the reading direction is left-to-right and
-  /// [scrollDirection] is [Axis.horizontal], then the scroll view scrolls from
-  /// left to right when [reverse] is false and from right to left when
-  /// [reverse] is true.
-  ///
-  /// Similarly, if [scrollDirection] is [Axis.vertical], then the scroll view
-  /// scrolls from top to bottom when [reverse] is false and from bottom to top
-  /// when [reverse] is true.
-  ///
   /// Defaults to false.
   final bool reverse;
 
@@ -63,75 +49,30 @@ class AutoAnimatedSliverList extends StatefulWidget {
 }
 
 class _AutoAnimatedSliverListState extends State<AutoAnimatedSliverList>
-    with
-        TickerProviderStateMixin<AutoAnimatedSliverList>,
-        ListAnimation<AutoAnimatedSliverList> {
-  Timer _timer;
+    with TickerProviderStateMixin<AutoAnimatedSliverList> {
+  final String _keyPrefix = utils.createCryptoRandomString();
 
-  @override
-  void initState() {
-    super.initState();
-    init();
-  }
-
-  void init({int from = 0}) {
-    itemsCount = from;
-    Future.delayed(widget.delay, () {
-      _timer = Timer.periodic(widget.showItemInterval, (Timer timer) {
-        if (itemsCount == widget.itemCount || !mounted) {
-          return timer.cancel();
-        }
-        insertItem(
-          widget.reverse ? 0 : itemsCount,
-          duration: widget.showItemDuration,
-        );
-      });
-    });
-  }
-
-  @override
-  void didUpdateWidget(AutoAnimatedSliverList oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (widget.itemCount < oldWidget.itemCount) {
-      init();
-    } else if (itemsCount < widget.itemCount && !(_timer?.isActive ?? true)) {
-      init(from: itemsCount);
-    }
-  }
-
-  @override
-  void dispose() {
-    _timer?.cancel();
-    for (ActiveItem item in incomingItems) {
-      item.controller.dispose();
-    }
-    for (ActiveItem item in outgoingItems) {
-      item.controller.dispose();
-    }
-    super.dispose();
-  }
-
-  Widget _itemBuilder(BuildContext context, int itemIndex) {
-    final ActiveItem outgoingItem = activeItemAt(outgoingItems, itemIndex);
-    if (outgoingItem != null) {
-      return outgoingItem.removedItemBuilder(
-          context, outgoingItem.controller.view);
-    }
-
-    final ActiveItem incomingItem = activeItemAt(incomingItems, itemIndex);
-    final Animation<double> animation =
-        incomingItem?.controller?.view ?? kAlwaysCompleteAnimation;
-    return widget.itemBuilder(context, indexToItemIndex(itemIndex), animation);
-  }
-
-  @override
-  Widget build(BuildContext context) => SliverList(
-        delegate: SliverChildBuilderDelegate(
-          _itemBuilder,
-          childCount: itemsCount,
+  Widget _itemBuilder(BuildContext context, int itemIndex) =>
+      AnimateOnVisibilityChange(
+        duration: widget.showItemDuration,
+        key: Key('$_keyPrefix.$itemIndex'),
+        builder: (context, animation) => widget.itemBuilder(
+          context,
+          itemIndex,
+          animation,
         ),
       );
 
   @override
-  TickerProvider get vsync => this;
+  Widget build(BuildContext context) => AnimateOnVisibilityWrapper(
+        delay: widget.delay,
+        showItemInterval: widget.showItemInterval,
+        useListStack: true,
+        child: SliverList(
+          delegate: SliverChildBuilderDelegate(
+            _itemBuilder,
+            childCount: widget.itemCount,
+          ),
+        ),
+      );
 }